fix 3 bugs: 1. need to flip image vertically after read the texture file; 2. in world2screen func, a very delicate difference makes a huge distinguish; 3. when calculating bary_centric, P can not be Vec3i, it should be Vec3f

+++ shader.h
@@ -29,4 +29,4 @@ public:
 
 void triangle(Vec3f* pts, BaseShader& shader, TGAImage& image, float* zbuffer, Model* model);
 
-Vec3f bary_centric(Vec2f A, Vec2f B, Vec2f C, Vec3i P);
+Vec3f bary_centric(Vec2f A, Vec2f B, Vec2f C, Vec3f P);
